+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.36.0] - 2021-02-24
 
 ### Added
 - Added set_return_format and get_return_format methods to Checkbox.
+
+### Fixed
+- A bug where RediPress queryable fields with array values may not have worked properly.
 
 ## [1.35.1] - 2021-01-27